feat: add tests and demo
