+++ sample/.eleventy.js
@@ -1,0 +1,5 @@
+const AddWebComponentDefinitions = require("../")
+
+module.exports = function(eleventyConfig) {
+	eleventyConfig.addPlugin(AddWebComponentDefinitions);
+};
