+++ sample/index.njk
@@ -1,0 +1,1 @@
+<custom-tag></custom-tag>
